python-selenium-pytest-allure/requirements.txt
allure-pytest==2.6.5
pytest==4.6.3
selenium==3.141.0